task3

+++ 0x0A-argc_argv/3-mul.c
@@ -9,7 +9,7 @@
  */
 int main(int argc, char *argv[])
 {
-	int i, j;
+	long int i, j;
 
 	if (argc != 3)
 	{
@@ -18,6 +18,6 @@ int main(int argc, char *argv[])
 	}
 	i = atoi(argv[1]);
 	j = atoi(argv[2]);
-	printf("%d\n", i * j);
+	printf("%ld\n", i * j);
 	return (0);
 }
